Refactored nvim config.
- All user configs (aside from ftplugin files) are now stored in the
  directory "lua/mattjamesdev".
- Moved all LSP config (server configs, cmp, lsp-installer) into their
  own directory "lua/mattjamesdev/plugin/lsp".
- Moved options config (i.e. "vim.opt") into its own file
  "lua/mattjamesdev/options.lua".
