+++ nvim/.config/nvim/lua/mattjamesdev/plugin/colourscheme.lua
