+++ nvim/.config/nvim/lua/mattjamesdev/plugin/init.lua
@@ -1,0 +1,14 @@
+require('mattjamesdev.plugin.lsp')
+require('mattjamesdev.plugin.efm-langserver')
+--require('mattjamesdev.plugin.null-ls')
+require('mattjamesdev.plugin.treesitter')
+require('mattjamesdev.plugin.lualine')
+require('mattjamesdev.plugin.nvim-tree')
+require('mattjamesdev.plugin.nvim-autopairs')
+require('mattjamesdev.plugin.toggleterm')
+require('mattjamesdev.plugin.bufferline')
+require('mattjamesdev.plugin.alpha-nvim')
+require('mattjamesdev.plugin.which-key')
+require('mattjamesdev.plugin.neodim')
+require('mattjamesdev.plugin.nvim-notify')
+require('mattjamesdev.plugin.colourscheme')
