+++ nvim/.config/nvim/lua/mattjamesdev/plugin/lsp/cmp.lua
@@ -1,9 +1,4 @@
---
--- LSP configurations
---
-
 vim.o.completeopt = 'menu,menuone,noselect'
-
 
 -- Configuration for nvim-cmp (completion)
 -- Setup nvim-cmp.
@@ -94,54 +89,3 @@ cmp.setup.cmdline(':', {
   }),
   mapping = cmp.mapping.preset.cmdline({})
 })
-
--------------------------------------------------------------------------------
--- Configuration for nvim-lsp-installer (copied from the GitHub page)
-local lsp_installer = require("nvim-lsp-installer")
-lsp_installer.setup({})
-
--------------------------------------------------------------------------------
--- Setup lspconfig.
-local capabilities = require('cmp_nvim_lsp').update_capabilities(vim.lsp.protocol.make_client_capabilities())
-
-local on_attach = function(client, bufnr)
-  -- Mappings.
-  -- See `:help vim.lsp.*` for documentation on any of the below functions
-  local bufopts = { noremap = true, silent = true, buffer = bufnr }
-  vim.keymap.set('n', 'gD', vim.lsp.buf.declaration, bufopts)
-  vim.keymap.set('n', 'gd', vim.lsp.buf.definition, bufopts)
-  vim.keymap.set('n', 'K', vim.lsp.buf.hover, bufopts)
-  vim.keymap.set('n', 'gi', vim.lsp.buf.implementation, bufopts)
-  vim.keymap.set('n', '<space>k', vim.lsp.buf.signature_help, bufopts)
-  vim.keymap.set('n', '<space>wa', vim.lsp.buf.add_workspace_folder, bufopts)
-  vim.keymap.set('n', '<space>wr', vim.lsp.buf.remove_workspace_folder, bufopts)
-  vim.keymap.set('n', '<space>wl', function()
-    print(vim.inspect(vim.lsp.buf.list_workspace_folders()))
-  end, bufopts)
-  vim.keymap.set('n', '<space>D', vim.lsp.buf.type_definition, bufopts)
-  vim.keymap.set('n', '<space>rn', vim.lsp.buf.rename, bufopts)
-  vim.keymap.set('n', '<space>ca', vim.lsp.buf.code_action, bufopts)
-  vim.keymap.set('n', 'gr', vim.lsp.buf.references, bufopts)
-  vim.keymap.set('n', '<space>lf', vim.lsp.buf.formatting, bufopts)
-end
-
-require('lspconfig').pyright.setup {
-  capabilities = capabilities,
-  on_attach = on_attach,
-}
-
-require('lspconfig').sumneko_lua.setup {
-  capabilities = capabilities,
-  on_attach = on_attach,
-  settings = {
-    Lua = {
-      telemetry = { enable = false },
-    }
-  }
-}
-
-require('lspconfig').tsserver.setup {
-  capabilities = capabilities,
-  on_attach = on_attach,
-  root_dir = function() return vim.loop.cwd() end,
-}
