+++ nvim/.config/nvim/lua/mattjamesdev/plugin/lsp/lsp-installer.lua
@@ -1,0 +1,3 @@
+-- Configuration for nvim-lsp-installer (copied from the GitHub page)
+local lsp_installer = require("nvim-lsp-installer")
+lsp_installer.setup({})
