+++ nvim/.config/nvim/lua/mattjamesdev/plugin/nvim-tree.lua
